+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/ColumnReferenceContributor.kt
@@ -3,11 +3,12 @@ package dev.ekvedaras.laravelquery.reference
 import com.intellij.patterns.PlatformPatterns
 import com.intellij.psi.PsiReferenceContributor
 import com.intellij.psi.PsiReferenceRegistrar
+import com.jetbrains.php.lang.psi.elements.impl.StringLiteralExpressionImpl
 
 class ColumnReferenceContributor : PsiReferenceContributor() {
     override fun registerReferenceProviders(registrar: PsiReferenceRegistrar) {
         registrar.registerReferenceProvider(
-            PlatformPatterns.psiElement(),
+            PlatformPatterns.psiElement(StringLiteralExpressionImpl::class.java),
             ColumnReferenceProvider()
         )
     }