+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/TableOrViewReferenceProvider.kt
@@ -7,6 +7,8 @@ import com.intellij.psi.PsiReference
 import com.intellij.psi.PsiReferenceProvider
 import com.intellij.util.ProcessingContext
 import com.jetbrains.php.lang.psi.elements.MethodReference
+import com.jetbrains.rd.util.addUnique
+import com.jetbrains.rd.util.lifetime.Lifetime
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderClassMethod
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderMethodByName
@@ -15,7 +17,15 @@ import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isTableParam
 import dev.ekvedaras.laravelquery.utils.MethodUtils
 
 class TableOrViewReferenceProvider : PsiReferenceProvider() {
+    companion object {
+        val isResolving = mutableListOf<PsiElement>()
+    }
+
     override fun getReferencesByElement(element: PsiElement, context: ProcessingContext): Array<PsiReference> {
+        if (isResolving.contains(element)) {
+            return PsiReference.EMPTY_ARRAY
+        }
+
         val method = MethodUtils.resolveMethodReference(element) ?: return PsiReference.EMPTY_ARRAY
         val project = method.project
 
@@ -23,11 +33,15 @@ class TableOrViewReferenceProvider : PsiReferenceProvider() {
             return PsiReference.EMPTY_ARRAY
         }
 
+        isResolving.addUnique(Lifetime.Eternal, element)
+
         val target = DbReferenceExpression(element, DbReferenceExpression.Companion.Type.Table)
         var references = arrayOf<PsiReference>()
 
         target.schema.parallelStream().forEach { references += SchemaPsiReference(target, it) }
         target.table.parallelStream().forEach { references += TableOrViewPsiReference(target, it) }
+
+        isResolving.remove(element)
 
         return references
     }